// web/src/api/generated/model/getOrgSlugReportsTransactions200ReportsPaidThisMonthTransactionsItem.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * HouseApp API
 * API for HouseApp
 * OpenAPI spec version: 1.0.1
 */
import type { GetOrgSlugReportsTransactions200ReportsPaidThisMonthTransactionsItemStatus } from "./getOrgSlugReportsTransactions200ReportsPaidThisMonthTransactionsItemStatus";

export type GetOrgSlugReportsTransactions200ReportsPaidThisMonthTransactionsItem =
  {
    id: string;
    title: string;
    amount: number;
    dueDate: string;
    /** @nullable */
    paidAt?: string | null;
    ownerName: string;
    ownerId?: string;
    payTo?: string;
    payToId?: string;
    payToName?: string;
    payToEmail?: string;
    status?: GetOrgSlugReportsTransactions200ReportsPaidThisMonthTransactionsItemStatus;
  };